pyproject.toml
[tool.poetry]
name = "rdb-lock-newbie"
version = "0.1.0"
description = ""
authors = ["Yuuki Takahashi <20282867+yktakaha4@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
mysql-connector-python = "^9.1.0"
psycopg = {extras = ["binary"], version = "^3.2.3"}


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
